chain-ethereum: cleanup types/utils
