+++ packages/chain-ethereum/src/types.ts
@@ -15,3 +15,15 @@ export type SIWEMessage = {
 	expirationTime: string | null
 	resources: string[]
 }
+
+export type EIP712SessionData = {
+	signature: Uint8Array
+}
+
+export type EIP712SessionMessage = {
+	address: string
+	publicKey: string
+	blockhash: string | null
+	timestamp: number
+	duration: number | null
+}
